requirements.txt
asn1crypto==0.24.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
cryptography==2.7
cycler==0.10.0
decorator==4.4.0
fpdf==1.7.2
future==0.17.1
idna==2.8
imageio==2.5.0
img2pdf==0.3.3
imutils==0.5.3
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
opencv-python==3.4.7.28
Pillow==6.1.0
pkg-resources==0.0.0
pycparser==2.19
pyparsing==2.4.2
python-dateutil==2.8.0
python-telegram-bot==12.1.0
PyWavelets==1.0.3
requests==2.22.0
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
tornado==6.0.3
urllib3==1.25.3
